fastapi/requirements.txt
fastapi
uvicorn[standard]

sqlmodel
sqlalchemy

psycopg2-binary

python-dotenv